transactions thing (#45)
* added transaction history (#22)

* added dashboard

* added AppDrawer component

* added transaction history

---------


* Database functions (#24)

* database functions

* initial function testing

* partial exception functions

---------


* image test working init on product upload

* still testing product image thingy

* save image post final

* post backend integration for products

* dropdown thing

* remove description prdocut

* needs refactor type shit and edit is gucci prlly

* image on edit products ting ug tung duplicate thing problem

* edit products image is now updatable

* transactions tentative

* transactions and shit

---------

+++ lib/models/ProductsModel.dart
@@ -20,8 +20,8 @@ class Product {
     return Product(
         id: userDoc,
         name: data['name'],
-        selling_price: data['selling_price'],
-        total_purchase: data['total_purchase'],
+        selling_price: data['selling_price'].toDouble(),
+        total_purchase: data['total_purchase'].toDouble(),
         product_stock: data['product_stock'],
         category: data['category'],
         image: data['image']);
